chore: djandorestframework installed
The djangorestframework was installed, and the requiremnts.txt was updated. A new Todo.viewset was implemented and a serializers.py was created for the todo model. The todo model and the rest_framework were registered in the settings.py. A router was defined/imported within the urls.py with the reference to the TodoViewset when the route is empty.