refactor: TodoViewSet added custom create method to TodoViewSet for handling Todo creation
Implemented a custom create method in the TodoViewSet to handle the creation of Todo objects. This method manually extracts the title and description from the POST request and assigns the currently authenticated user to the new Todo. The response is serialized as JSON and returned as an HttpResponse. This approach allows for more customized processing during the creation of Todo items.